wandb-weave-example/pyproject.toml
[tool.poetry]
name = "wandb-weave-example"
version = "0.1.0"
description = ""
authors = ["Uragami Taichi <backpaper0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.16"
langchain-openai = "^0.1.24"
python-dotenv = "^1.0.1"
weave = "^0.51.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
